算術演算子関連の関数を本体から分離した
